LayoutTests/platform/ios-simulator/editing/selection/doubleclick-whitespace-expected.txt

This tests that double-clicking a word on the Windows platform selects the whitespace after the word.

Doubleclickme     |END|
Doubleclickme|END|
Doubleclickme    |END|
Doubleclickme
   |(Should not have been extended into this line)|
Doubleclickme      
Doubleclickme   			 |END|

Failed test1
  Expected: Doubleclickme     
  Actual: 
Failed test2
  Expected: Doubleclickme
  Actual: 
Failed test3
  Expected: Doubleclickme    
  Actual: 
Failed test4
  Expected: Doubleclickme
  Actual: 
Failed test5
  Expected: Doubleclickme    
  Actual: 
Failed test6
  Expected: Doubleclickme   			 
  Actual: 
